+++ src/Resources/app/administration/src/main.js
@@ -3,6 +3,7 @@ import './module/sw-cms';
 import './module/system-config';
 import './component/api-test-button';
 import './component/freshdesk-agent-select';
+import './component/freshdesk-custom-field-select';
 import './component/freshdesk-group-select';
 import './component/freshdesk-product-select';
 import './component/freshdesk-company-select';